main.py
You enter the date you are interested in and color of the markers on the map.
The program will return all the events which are held on this date and their number. 
It will also create "Map.html" with all the names of events as  hypertext to their webpage on eventful.